pyproject.toml
[project]
name = "challenge-list-notes"
version = "0.1.0"
description = "Project to create and list notes"
authors = [
    {name = "Ana Elles",email = "alelles16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
